Modification: resolved a bug which can't delete the generated doc after doing the "Empty recycle bin" action.
git-svn-id: http://talendforge.org/svn/tos/trunk@12112 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.repository/src/main/java/org/talend/repository/ui/actions/EmptyRecycleBinAction.java
@@ -22,6 +22,7 @@ import org.eclipse.ui.IWorkbenchPage;
 import org.eclipse.ui.PlatformUI;
 import org.talend.commons.exception.MessageBoxExceptionHandler;
 import org.talend.commons.ui.image.ImageProvider;
+import org.talend.core.model.repository.ERepositoryObjectType;
 import org.talend.core.model.repository.IRepositoryObject;
 import org.talend.core.ui.images.ECoreImage;
 import org.talend.repository.i18n.Messages;
@@ -84,8 +85,9 @@ public class EmptyRecycleBinAction extends AContextualAction {
                                 page.closeEditor(editors.getEditor(false), false);
                             }
                         }
-
-                        factory.deleteObjectPhysical(objToDelete);
+                        if (objToDelete.getType() != ERepositoryObjectType.JOB_DOC
+                                && objToDelete.getType() != ERepositoryObjectType.JOBLET_DOC)
+                            factory.deleteObjectPhysical(objToDelete);
                     }
                 }
             } catch (Exception e) {
